+++ sdk/spring/azure-spring-boot-starter-active-directory/CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## 3.7.0-beta.1 (Unreleased)
 ### New Features
-- Add property `azure.activedirectory.application-type` to explicitly configure the application scenario.
+- Add AAD property `azure.activedirectory.application-type` to explicitly configure the application scenario.
   The following 4 values can be configured:
   - *web_application*: Not required, only Web Application scenario.
   - *resource_server*: Not required, only Resource Server scenario.
-  - *resource_server_with_obo*: Not required, Resource Server with authorization grant type 'on-behalf-of' scenario.
+  - *resource_server_with_obo*: Not required, Resource Server with authorization grant type 'on_behalf_of' scenario.
   - *web_application_and_resource_server*: Required, Web Application and Resource Server are in the same application.
+- Support new value 'on_behalf_of' for AAD Obo authorization grant type.
 
 ## 3.6.0 (2021-06-23)
 ### Breaking Changes